# Conflicts:
#	admin/src/main/java/com/exadel/frs/service/ModelService.java
#	admin/src/test/java/com/exadel/frs/ModelServiceTest.java
#	admin/src/test/java/com/exadel/frs/OrganizationServiceTest.java

+++ api/src/main/java/com/exadel/frs/core/trainservice/component/FaceClassifierPredictor.java
@@ -19,8 +19,8 @@ package com.exadel.frs.core.trainservice.component;
 import com.exadel.frs.core.trainservice.dao.ModelDao;
 import lombok.RequiredArgsConstructor;
 import lombok.val;
-import org.springframework.context.ApplicationContext;
 import org.apache.commons.lang3.tuple.Pair;
+import org.springframework.context.ApplicationContext;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -30,7 +30,7 @@ public class FaceClassifierPredictor {
     private final ModelDao modelDao;
     private final ApplicationContext context;
 
-    public Pair<Integer, String> predict(final String modelKey, final double[] input) {
+    public Pair<Double, String> predict(final String modelKey, final double[] input) {
         val model = modelDao.getModel(modelKey);
         val faceClassifier = context.getBean(FaceClassifierAdapter.class);
         faceClassifier.setClassifier(model);
